requirements.txt
opencv-python>=4.8.0
dlib>=19.24.0
numpy>=1.24.0
scipy>=1.10.0
pyserial>=3.5
imutils>=0.5.4
